requirements.txt
langchain
langgraph
autogen
requests
python-dotenv
gradio
streamlit
pandas
pytest
pytest-cov
tenacity
aiohttp
huggingface_hub
groq
graphviz
numpy
langchain-core
langchain-community
langsmith>=0.1.75
rich==13.7.1
fastapi==0.95.2
uvicorn[standard]==0.23.2
streamlit==1.28.0
fpdf
pydot
networkx
matplotlib
